Put navbar code back into the templates.
I put the navbar back in instead of having it in a fragment so that the correct link can be set to active when navigating the site. There is a problem where the Admin dropdown won't work after navigating to one of the admin pages unless you navigate to the Home tab first.